Implemented several models methods
Implemented the Level#completed?(user_id) method, which checks for user achievements for all the level topics
Implemented the Topic#completed?(user_id) method, which checks if the user has completed all the topic contents
Implemented the Content#completed?(user_id) method, which checks if the user has solved attempts for all the topic questions
Implemented the Question#answered?(user_id) method, which checks if the user has a solved attempt for the question
Started the implementation of Attempt#create_attempt(user_id,question_id,option_id). Now facing some design choices about using Mongoid Listeners in order to ensure the check of the business rules or the game every time an attempt is created, like checking if an achievement needs to be created as well.