Started implementation of Integration Tests with Capybara.
Refactored controller_macros. Now the module is called "login_helpers", and includes 2 login methods to be used by controller and integration tests. The new login_with method was necessary because the Devise sign_in helpers are not avaliable inside the integration tests scope.
With whis, it was necessary do include in the spec_helper that the module is both a controller, request and feature module. For some reason, the before(:each) hook inside the login method stopped working. I removed it from inside the method and added in en each admin controller spec file. Now the tests are working again.
User show page now shows the achievements for that user.